fix(CRUD): Se arreglo el crud para editar/eliminar/crear semanas.

+++ front/src/assets/services/deleteWeek.js
@@ -1,11 +1,22 @@
 import { BASEURL } from "../db/connectURL";
+import { obtenerDelSessionStorage } from "../utils/localStorage";
 
 export async function handleDeleteWeek(id, onSemanasUpdate) {
+  const token = obtenerDelSessionStorage("token");
+
+  if (!token) {
+    console.error(
+      "Token no encontrado. Debes iniciar sesion como administrador para poder hacer peticiones"
+    );
+    return false;
+  }
+
   try {
     const response = await fetch(`${BASEURL}/api/semanas/${id}`, {
       method: "DELETE",
       headers: {
         "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
       },
     });
